Less aggressive refactoring
